+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 * **Breaking change**: Changed positional arguments of
   [HexaPDF::Layout::TextBox::new] and [HexaPDF::Layout::ImageBox::new] to
   keyword arguments for a consistent box initialization interface
+* Layout box implementations to provide a `#supports_position_flow?` method that
+  indicates whether the box supports flowing its content around other content.
 
 ### Fixed